+++ temporalio/worker/_worker.py
@@ -129,6 +129,7 @@ class Worker:
         default_heartbeat_throttle_interval: timedelta = timedelta(seconds=30),
         max_activities_per_second: float | None = None,
         max_task_queue_activities_per_second: float | None = None,
+        max_eager_activity_reservations_per_workflow_task: int = 3,
         graceful_shutdown_timeout: timedelta = timedelta(),
         workflow_failure_exception_types: Sequence[type[BaseException]] = [],
         shared_state_manager: SharedStateManager | None = None,
@@ -267,6 +268,10 @@ class Worker:
                 poll request. If multiple workers on the same queue have
                 different values set, they will thrash with the last poller
                 winning.
+            max_eager_activity_reservations_per_workflow_task: Maximum number of
+                activity slots that may be reserved for eager execution when
+                completing a workflow task. The default is 3. Setting this to
+                zero disables eager activity execution.
             graceful_shutdown_timeout: Amount of time after shutdown is called
                 that activities are given to complete before their tasks are
                 cancelled.
@@ -369,6 +374,7 @@ class Worker:
             default_heartbeat_throttle_interval=default_heartbeat_throttle_interval,
             max_activities_per_second=max_activities_per_second,
             max_task_queue_activities_per_second=max_task_queue_activities_per_second,
+            max_eager_activity_reservations_per_workflow_task=max_eager_activity_reservations_per_workflow_task,
             graceful_shutdown_timeout=graceful_shutdown_timeout,
             workflow_failure_exception_types=workflow_failure_exception_types,
             shared_state_manager=shared_state_manager,
@@ -656,6 +662,9 @@ class Worker:
                 max_activities_per_second=config.get("max_activities_per_second"),
                 max_task_queue_activities_per_second=config[
                     "max_task_queue_activities_per_second"
+                ],  # type: ignore[reportTypedDictNotRequiredAccess]
+                max_eager_activity_reservations_per_workflow_task=config[
+                    "max_eager_activity_reservations_per_workflow_task"
                 ],  # type: ignore[reportTypedDictNotRequiredAccess]
                 graceful_shutdown_period_millis=int(
                     1000 * config["graceful_shutdown_timeout"].total_seconds()  # type: ignore[reportTypedDictNotRequiredAccess]
@@ -977,6 +986,7 @@ class WorkerConfig(TypedDict, total=False):
     default_heartbeat_throttle_interval: timedelta
     max_activities_per_second: float | None
     max_task_queue_activities_per_second: float | None
+    max_eager_activity_reservations_per_workflow_task: int
     graceful_shutdown_timeout: timedelta
     workflow_failure_exception_types: Sequence[type[BaseException]]
     shared_state_manager: SharedStateManager | None